% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/data_submission.R
\name{NGS_data_submission}
\alias{NGS_data_submission}
\title{Data submission function}
\usage{
NGS_data_submission(print = TRUE)
}
\arguments{
\item{input}{.csv community and environmental matrix}
}
\description{
This function can be used for data submission of file type .csv. It reads a community data matrix (site by species), a contextual data matrix and a biome (user defined
factors) matrix. The submission uses file.choose() and uses the headers and row.names of the .csv file.
After submission of the data it give total read abundace per sample. Additionally it makes a scatterplot of the contextual data based on a log transformation
}
\examples{
NGS_data_submission()
}
\keyword{biome}
\keyword{comm,}
\keyword{contex,}

