% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/diveristy_analysis.R
\name{NGS_diversity_indecies_function}
\alias{NGS_diversity_indecies_function}
\title{Diveristy analysis of NGS data}
\usage{
NGS_diversity_indecies_function(x, print = TRUE)
}
\arguments{
\item{normalised}{community matrix, print default =TRUE}
}
\description{
This function anaylsis a normalised community matrix using Vegan based functions. It requires a species by site matrix and analyses this for species RICHNESS, EVENESS and DIVERSITY INDECIES ANALYSIS
- Does species analysis :richness
- Does species frequency analysis (freuqency of occurance)
- Does Shannon Weaver analysis
- Does Simpson analysis
- Does Inverse Simpsoms analysis
- Does Evenness analysis
- All results are stored in file diversity_results.csv, all graphs (.png images) are stored
  in the working directory. They are high quailty images.
}
\examples{
NGS_diversity_index_function(x, print=TRUE)
}
\keyword{analysis}
\keyword{diveristy}

