% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/NMDS_analysis.R
\name{NGS_NMDS_function}
\alias{NGS_NMDS_function}
\title{NMDS Analysis}
\usage{
NGS_NMDS_function(x, y, z, plot = TRUE, sites = TRUE)
}
\arguments{
\item{normalised}{community matrix, default plot=TRUE, default site=TRUE}
}
\description{
This function anaylsis a normalised community matrix, contextual matrix and factor matrix
- It requires a species by site matrix, contextual data matrix and selection factors matrix
- It calculates a distance matrix using vegdist() by bray curtis and jaccard dissimilarity analysis.
- Then it uses the metaMDS() function to perform nonmetric multidimensional scaling (NMDS).
- Subsequently it prints a stress plot to assess goodness of the ordination fit
- Then it plots a basic NMDS, this can be done with "sites" or "species" labeled.
- Subsequently more complex NMDS plots can be made with seperations or coloring based on user defined factors or contextual data.
- Then the funtion does rankindex() of community and contex data and analysis the significant of environmental factors on the community dissimilarity using the envfit() function.
- Results are listed after analysis
}
\examples{
NGS_NMDS_function(x,y,z, plot=FALSE)
}
\keyword{NMDS}
\keyword{analysis}

