% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/normalisation.R
\name{NGS_normalise_abundance_function}
\alias{NGS_normalise_abundance_function}
\title{Normalisation of community data}
\usage{
NGS_normalise_abundance_function(x)
}
\arguments{
\item{input}{.csv community or environmental matrix}
}
\description{
This function can be used to normalise both community and contextual data using the vegan decostand () function. Choose the method of normailsation wisely !!
Its exports the normalised community matrix and a normalisation "checked" matrix.
}
\examples{
NGS_normalise_abundance_function(x)
}
\keyword{normalisation}

