% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/significance_functions.R
\name{NGS_Significance_function}
\alias{NGS_Significance_function}
\title{NMDS Analysis}
\usage{
NGS_Significance_function(x, y)
}
\arguments{
\item{normalised}{community matrix, default plot=FALSE}
}
\description{
This function tests whether there is a satistically significant difference between two or more groups of sampling units (community and contextual data)
- The function uses normalised community matrix, then does a dissimilartiy analysis using VEGAN vegdist() using the Jaccard index and a subsequent metaMDS() analysis.
- It subsequently uses the VEGAN envfit() analysis on a contextual data matrix.
- Finally significance test can be done using anosim() and mantel() functions.
}
\examples{
NGS_Cluster_analysis(x)
}
\keyword{NMDS}
\keyword{analysis}

