% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/NGS_grep.R
\name{NGSgrep}
\alias{NGSgrep}
\title{NGSgrep}
\usage{
NGSgrep(x)
}
\arguments{
\item{x,}{vector}
}
\description{
This function allows you to grep a specific string from a column name. For example a specific taxnomic name or groups. All the columns containing the string and their associated values are copied into a new vector.
}
\examples{
NGSgrep(x)
}
\keyword{grep}

